Pull Automaton weight out so it can be shared
